+++ README.md
@@ -9,20 +9,20 @@ tested with **Python 2.7.6** and **Python 3.4.0**
 ### Instalation
 
 ```sh
-pip install nominatim
+pip install pynominatim
 ```
 
 or
 
 ```sh
-git clone https://github.com/damianbraun/nominatim.git
+git clone https://github.com/Xevib/pynominatim.git
 python setup.py install
 ```
 
 ### Quick Start
 
 ```
-In [1]: from nominatim import Nominatim, NominatimReverse
+In [1]: from pynominatim import Nominatim, NominatimReverse
 
 In [2]: nom = Nominatim()
