+++ pynominatim/__init__.py
@@ -1,4 +1,4 @@
 """Nominatim API
 http://wiki.openstreetmap.org/wiki/Nominatim"""
 from __future__ import absolute_import
-from .nominatim import Nominatim, NominatimReverse, NominatimException
+from .pynominatim import Nominatim, NominatimReverse, NominatimException
